+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 - `Bernoulli` distribution is now generic over the floating point and unsigned data types
 - `Beta` distribution is now generic over the floating point data types
 - `Chi` distribution is now generic over the floating point data types
+- `ChiSquared` distribution is now generic over the floating point data types
 - `DiscreteUniform` distribution is now generic over the the floating point and signed data types
 - `Exponential` distribution is now generic over the floating point data types
+- `Gamma` distribution is now generic over the floating point data types
 - `Poisson` uses `u64` instead of `i64` now
 
 ### v0.3.1